Refactor: NetworkModule, Qualifier 파일 경로 변경
